sem: The SEM Function

Structural Equation Models (SEM) and particular cases via rstan interface

sem(
  data,
  paths,
  blocks,
  signals,
  row_names = rownames(data),
  prior_specs = list(coef = c("normal(0,1)"), error_var = c("inv_gamma(2.1, 1.1)")),
  cores = parallel::detectCores(),
  pars = c("alpha", "lambda", "sigma2", "Xna"),
  iter = 2000,
  chains = 4,
  scaled = TRUE,
  ...
)

Arguments

data
a mandatory data.frame object where the columns are variables and the rows are observations

paths
list referring to the inner model

blocks
list of column names (or integers in 1:ncol(data)) indicating the manisfest variables correpoding to each block;

signals
list referring to the signals of the factor loadings initial values; must be true: (length(signals) == length(blocks)) && (lengths(signals) == lengths(blocks))

prior_specs
prior settings for the Bayesian approach; only `normal(0,...)` for coef and `inv_gamma` for error_var are temporarily available, beta is ignored if paths are not specific; next versions should include more options

cores
number of core threads to be used

pars
which parameters to be included in the outcome; options are any subset of default c("alpha", "lambda", "sigma2", "Xna")

iter
number of iterations

chains
number of posterior chains

scaled
logical; indicates whether to center and scale the data

...
further arguments passed to or from other methods

rownames
refers to the observation vector of characters; must have nrow(data) lenght; defaults to rownames(data)

Value

An object of class sem; a list of 14 to 16:

stanfit

S4 object of class stanfit

posterior

the list of posterior draws separate by chains

model

character; pointer to pre-defined stan model

mean_loadings

matrix of factor loadings posterior means

mean_scores

matrix of factor scores posterior means

mean_var

vector of error variances posterior means

coef

vector of regression coefficients posterior means

stats

posterior descriptives statistics

blocks

list of blocks

paths

list of paths

credint

Highest posterior density intervals (HPD)

h

vector of communalities

PVTE

vector of total variance proportions

AFR2

adjusted coefficient of determination

SQE

explained sums of squares

SQT

total sums of squares

Details

Fits the SEM to specific data

Examples

data('set1')

fit1 <- sem(data = set1$set, blocks = set1$blocks,
 chains = 1)

#> 
#> SAMPLING FOR MODEL 'factorial' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000831 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 8.31 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 14.12 seconds (Warm-up)
#> Chain 1:                6.46532 seconds (Sampling)
#> Chain 1:                20.5854 seconds (Total)
#> Chain 1: 
#> 
#> Extracting posterior samples, please wait...
#> 
#> Done!

summary(fit1)

#> 
#> 
#> ---
#> bsem model:  factorial 
#> latent variables (outter model):  3 
#> 
#> 
#>  outer model loadings:
#> 
#>  latent1 
#>         mean se_mean     sd    2.5%   97.5% n_eff  Rhat
#> var1 -0.9302  0.0126 0.0978 -1.1386 -0.7574    62 1.011
#> var2 -0.8974  0.0121 0.0970 -1.1039 -0.7093    66 1.013
#> var3  0.8018  0.0111 0.1053  0.6197  1.0243    92 1.006
#> 
#>  latent2 
#>         mean se_mean     sd   2.5%   97.5% n_eff  Rhat
#> var4 -1.0453  0.0119 0.0841 -1.221 -0.8830    58 1.050
#> var5 -0.9895  0.0127 0.0864 -1.162 -0.8201    67 1.052
#> 
#>  latent3 
#>         mean se_mean     sd   2.5%   97.5% n_eff   Rhat
#> var6 -0.9254  0.0075 0.0930 -1.116 -0.7576   154 0.9992
#> var7 -0.9933  0.0077 0.0862 -1.172 -0.8336   123 1.0005
#> 
#> ---
#> 
#>  variances:
#>           mean se_mean     sd   2.5%  97.5% n_eff   Rhat
#> var[1]  0.2885  0.0015 0.0434 0.2163 0.3864   833 1.0063
#> var[2]  0.3416  0.0017 0.0515 0.2525 0.4524   892 0.9999
#> var[3]  0.4798  0.0025 0.0728 0.3610 0.6431   845 1.0026
#> var[4]  0.1082  0.0007 0.0187 0.0784 0.1492   732 1.0001
#> var[5]  0.2002  0.0009 0.0301 0.1493 0.2661  1002 1.0018
#> var[6]  0.2842  0.0016 0.0477 0.2088 0.3912   815 0.9992
#> var[7]  0.1782  0.0016 0.0355 0.1176 0.2565   499 0.9997
#> var[8]  0.6361  0.0023 0.0915 0.4830 0.8377  1549 1.0017
#> var[9]  0.1019  0.0006 0.0175 0.0730 0.1394   823 1.0015
#> var[10] 0.2052  0.0013 0.0347 0.1459 0.2794   696 0.9995
#> var[11] 0.3310  0.0015 0.0512 0.2429 0.4406  1222 0.9995
#> var[12] 0.0805  0.0005 0.0150 0.0570 0.1167  1007 0.9994
#> var[13] 0.1827  0.0010 0.0322 0.1310 0.2560   949 0.9995
#> var[14] 0.1749  0.0013 0.0321 0.1240 0.2450   623 1.0024
#> var[15] 0.2391  0.0012 0.0389 0.1743 0.3314  1106 0.9992
#> var[16] 0.6571  0.0026 0.0949 0.4897 0.8628  1384 1.0022
#> var[17] 0.3189  0.0015 0.0462 0.2410 0.4249   902 0.9993
#> var[18] 0.1453  0.0007 0.0234 0.1043 0.1995  1097 1.0022
#> var[19] 0.3185  0.0015 0.0483 0.2364 0.4260  1076 1.0005
#> var[20] 0.0732  0.0004 0.0133 0.0522 0.1045   992 1.0005
#> 
#> 
#> 
#> Head Rhat:
#> alpha[11,2] alpha[10,2]  alpha[5,2]  alpha[4,2] alpha[19,2] alpha[12,2] 
#>       1.060       1.053       1.052       1.050       1.043       1.039 
#> 
#> Tail Rhat:
#> lambda[3,29] lambda[1,99] lambda[1,93] lambda[3,94] lambda[1,10] lambda[2,12] 
#>        0.999        0.999        0.999        0.999        0.999        0.999 
#> 
#> median PVTE:  80.62 
#> 
#> ---

if (FALSE) {
data('set2')

fit2 <- sem(data = set2$set, blocks = set2$blocks,
 chains = 4, iter = 2500, warmup = 1000)
summary(fit2)
}